Arregla bugs en algo

+++ execute.h
@@ -2,6 +2,6 @@
 #define AC_CONTROL_H
 
 // Function to set AC temperature, humidity, and presence
-const char *execute(int temperature, int humidity, int presence);
+const char *execute(float temperature, float humidity, int presence);
 
 #endif
